+++ src/diagnostic/io/SpecFile.h
@@ -16,13 +16,13 @@ public:
 	inline QVector<QString>& channelNames() { return mChannelNames; }
 
 	const QString& name() const override { return mName; }
-	size_t width() const override { return mWidth; }
-	size_t height() const override { return mHeight; }
-	size_t channelCount() const override { return mChannelNames.size(); }
-	size_t viewChannelCount() const override { return 1; }
-	const QString& channelName(size_t c) const override { return mChannelNames.at(c); }
+	int width() const override { return mWidth; }
+	int height() const override { return mHeight; }
+	int channelCount() const override { return mChannelNames.size(); }
+	int viewChannelCount() const override { return 1; }
+	const QString& channelName(int c) const override { return mChannelNames.at(c); }
 
-	float value(size_t x, size_t y, size_t channel) const override
+	float value(int x, int y, int channel) const override
 	{
 		return mData[y * mWidth * channelCount() + x * channelCount() + channel];
 	}
@@ -32,6 +32,6 @@ private:
 	QVector<float> mData;
 	QVector<QString> mChannelNames;
 
-	size_t mWidth;
-	size_t mHeight;
+	int mWidth;
+	int mHeight;
 };
